+++ package.json
@@ -18,6 +18,7 @@
     "@docusaurus/core": "2.4.0",
     "@docusaurus/preset-classic": "2.4.0",
     "@mdx-js/react": "^1.6.22",
+    "clipboard-copy": "^4.0.1",
     "clsx": "^1.2.1",
     "hast-util-is-element": "^1.1.0",
     "katex": "^0.16.7",
